% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{fnll}
\alias{fnll}
\title{fonction RTMB qui devra être optimisée}
\usage{
fnll(param, fit = TRUE)
}
\arguments{
\item{param}{les parametres à optimiser}

\item{fit}{si la fonction retour nll ou uniquement les résultat de parametre initiaux}
}
\value{
nll
}
\description{
fonction RTMB qui devra être optimisée
}
\examples{
NULL
}
